feat: Devide Header, Footer component
