+++ components/Intro.tsx
@@ -2,7 +2,7 @@
 
 export default function Intro(): JSX.Element {
   return (
-    <section className=" text-white bg-black h-[calc(100vh-88px)]">
+    <section id="INTRO" className=" text-white bg-black h-[calc(100vh-88px)]">
       <div className="flex flex-col items-center justify-center w-11/12 h-full gap-4 m-auto">
         <p className="text-2xl font-bold">{"Hello I'm James An."}</p>
         <div className="flex justify-center gap-4">
